5-evaluation/requirements.txt
openai-agents>=0.2.0
composio-openai-agents>=0.5.0
python-dotenv>=1.0.0
openai>=1.0.0
asyncio
dataclasses
datetime
typing 